move classify_species file to services folder
